n_reclaimed stats accouting remove atomic stats.
Use the latch protection instead of the atomic stats accounting,
on the write path latch protection is already there, now we are
adding it to maintanence APIs like footprint() and to_stats(),
which are neither frequent nor latency critical.

Add n_reclaimed stats accounting that maintains a book-keeping
of reclaimed nodes.