mvcc: Implement a chain of mvcc snapshot from latest to oldest.
The older implementation uses a mvcc snapshot chain from oldest
to latest.

In the former method, snapshots are automatically dropped when
the last reference to the snapshot, typically a read operation,
exits the snapshot.

In this current method snapshots are to be cleaned by the
writer thread. While in the former method alloc/free can happen
in different threads, here they are done in the writer thread.

This patch is actually happens on SHA fab1d807b48bc49f43f3602e9d79b2070d31e6b3