robt: handle compaction leading to empty index.
There can be a corner case where compaction's cutoff removes
all the entry in the Robt index. In such cases we need to
preserve the last-seqno to give consistent view to application.

This simply means, compact operation only removes older entries
and it _does not_ rewind the index to initial state.

Consequently, fix the validation logic that check for seqno validity.